Add some new assertion helpers
`assert_flash_unit(options)`: Asserts that the `flash` method contains a
flash message in a given type slot. `options` must include a `type`
parameter, with a symbol representing the type of flash message (:success,
:danger etc). `options` may also include an `expected` parameter, with a
string value that the selected flash message must contain. Only for
use in Unit tests.

`assert_flash_integration(options)`: Asserts that a flash message is
rendered on the page. `options` may include a `type` parameter, with a
string representing the alert class module (e.g. passing 'success' will
look for an element with the `.alert-success` class). `options` may also
include an `expected` parameter with a string value that the flash
message must contain. Only for use in integration tests.

`assert_flash(options)`: Asserts that a flash message exists. Calls
`assert_flash_unit` or `assert_flash_integration`, depending on the type
of test currently being run. Passes it's `options` parameter directly to
the helper it ultimately calls. It is expected that the author knows
which `options` format to use.