Add some more assertion helpers
`assert_highlights(highlights)`: Asserts that a set of passed selectors
matches elements that are contained within an element with the class
`field_with_errors`. Such fields will be highlighted in red to indicate
errors. `highlights` must be an array, and it should contain strings
representing css selectors for form fields expected to be highlighted.
The assertion will fail if any specified fields are not highlighted, but
it will pass if additional unspecified fields are highlighted. This has
only been tested in integration tests, but it should work for
unit tests as well.

`assert_explanations(explanations)`: Asserts that a set of error
explanations exist on the page.  An error explanation is an `<li>`
element inside the element with id `error_explanation`.  These elements
appear when form validation fails.  `explanations` must be an array, and
it should contain strings identical to expected error explanations.  The
assertion will fail if a specified explanation is not found, but it will
pass if additional unspecified explanations exist.  This has only been
tested in integration tests, but it should work for unit tests as well.

`assert_error_messages(options)`: A high-level wrapper for
`assert_flash`, `assert_highlights`, and `assert_explanations`.
`options` should contain options hashes for the required assertion
helpers: `options[:highlights]` may be an options hash for
`assert_highlights`, `options[:explanations]` may be an options hash for
`assert_explanations`, and `options[:flash]` may be an options hash for
`assert_flash`.  If any of these parameters are omitted, the matching
assertion helper will not be called.  It is up to the author to make
sure that valid option hashes are included in `options`.  This assertion
will only fail if any of the consequence assertions fail, and will pass
if none are triggered (e.g. if an empty `options` has is passed).  This
has only been tested in integration tests, but it should work for unit
tests as well.