Make `assert_flash` accept `false` for `expected`
This allows assert_flash to be used to assert that a given flash type
doesn't exist. `false` can be passed in the `expected` options parameter
to do so.

Also upped the maximum method length to 15 lines for test files, because
some of these assertion helpers are just too complex for 10 lines