Make signup result helpers
`assert_signup_successful(name)`: Asserts that the user profile page has
been rendered, there is no `danger` flash, there is a `success` flash,
and the created user is logged in. The `name` parameter is optional, and
it must be a string representing the new user's `name` if it is passed.
If such a parameter is passed, the `success` flash will be asserted to
contain a personalized welcome message.

`assert_signup_failed(errors)`: Asserts that the user signup page has been
rendered, that a user is not logged in, and optionally, that a given set
of error messages exist.  The `errors` parameter will be passed directly
to `assert_error_messages` as it's `options` parameter. This assertion
will fail if any of the aforementioned assertions fail.