Move the user entry view into a partial.
I could have just done `render @users` instead of explicitly looping
over it, but that's too magic for my taste.  Even this is a little
mysterious: How does rails know to render the `app/views/users/_user`
template for the `user` variable? I suppose it knows by looking at the
type, and inferring a location, but that's the outer limit of what I
consider acceptably magic. I might refactor it to be more explicit later
on anyways.