trigger account activation email on signup
Instead of logging in, we send the activation email.  Moar security, and
knowledge that emails are verified is good.  In the future, I'd want to
have a richer verification state system, so users can sign up and use
the app right away, but have to verify their email before doing anything
important.