Fix bug in user index test re: unverified users
The test was looking for the first 3 users, regardless of if they are
verified, but I'm making it so that only admins can see unverified
users.  My unerified fixture (crichton) was showing up in the first 30,
so the test would be looking for him.  Now, the test specifically looks
for the first 30 verified users.