simplified the `correct_user_or_go_home` helper
Now it just redirects to home if the wrong user is logged in.  Much
simpler, and more semantically named.

I also removed the 'not logged in' test case.  While technically, this
constitutes a gap in test coverage, this particular case is a) hard to
test, and b) handled just fine anyway.  The app will never have an
action protected by `correct_user_or_go_home` but NOT `logged_in_user`
(or at least that would be very weird if it did), so this case is
handled by `logged_in_user` anyways.  In addition, were that to occur,
it would still redirect to home saying 'permission denied' - the only
'bug' would be that it doesnt render the login page right away.  Hardly
a serious problem, in my opinion.