Move the new micropost object into template
This saves me from having to instance it when I just render templates.
Since it's just creating a blank object for the current user, I don't
see any real risk here.  It might be a little strange to embed that
particular line of code in the template, but it solves some problems, so
I'm going to try it.