reorganize a couple functions in layout
I like to keep my functions defines before they are used, but at the
last minute. This keeps stuff well organized and encapsulated.  I'm
considering doing it the other way around though - defined immediately
after they are used the first time.  This is a more natural way to read
the code, and the order of definition doesn't matter, really