add `assert_at_page`
To be used in step functions to make sure the step is being used
correctly.  This could also be used to check the current path, but it's
better to check for know elements on the page, since the user doesn't
care about the path - they care about what they see.